+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Changelog
 
-### [unreleased]
+### KiteViewers v0.4.11 - 2024-04-05
 
 #### Added
 - add menu time_lapse for changing the simulation/ replay speed
+- add field `mod_text` to Viewer3D; it defines the quotient of the update rate and the text update rate
 
 ### KiteViewers v0.4.10 - 2024-03-30